requirements.txt
psycopg==3.2.9
tzdata==2025.2
